+++ examples/get_load_factors_batch.py
@@ -27,9 +27,7 @@ def main():
         lossesAvailability=0.1,
         lossesWake=0,
         numberOfTurbines=12,
-        roughnessLength=0.02,
-        usePowerCurveSmoothing=False,
-        useReanalysisCorrection=False,
+        usePowerCurveSmoothing=False
     )
 
     list_of_parameters = [