+++ tests/test_load_factors.py
@@ -17,7 +17,7 @@ def amun_session():
 def test_era5(amun_session):
     log.info("Testing for era5: expect correct successful response with load factors")
 
-    flow_parameters = BuiltInWindParameters("era5")
+    flow_parameters = BuiltInWindParameters("era5", True)
     base_parameters = LoadFactorBaseParameters(
         turbineModelId=amun_session.get_turbine_by_name("Siemens SWT-4.0-130")["id"],
         latitude=59.59,
@@ -29,9 +29,7 @@ def test_era5(amun_session):
         lossesAvailability=0.1,
         lossesWake=0,
         numberOfTurbines=12,
-        roughnessLength=0.02,
-        usePowerCurveSmoothing=False,
-        useReanalysisCorrection=True,
+        usePowerCurveSmoothing=False
     )
     result = amun_session.run_load_factor_for_parameters(flow_parameters, base_parameters)
     assert isinstance(result, dict)
@@ -44,7 +42,7 @@ def test_era5(amun_session):
 def test_era5_with_bad_base_parameters(amun_session):
     log.info("Testing for era5: expect error due to unsupported region")
 
-    flow_parameters = BuiltInWindParameters("era5")
+    flow_parameters = BuiltInWindParameters("era5", True)
     base_parameters = LoadFactorBaseParameters(
         turbineModelId=amun_session.get_turbine_by_name("Siemens SWT-4.0-130")["id"],
         latitude=159.59,
@@ -56,9 +54,7 @@ def test_era5_with_bad_base_parameters(amun_session):
         lossesAvailability=0.1,
         lossesWake=0,
         numberOfTurbines=12,
-        roughnessLength=0.02,
-        usePowerCurveSmoothing=False,
-        useReanalysisCorrection=True,
+        usePowerCurveSmoothing=False
     )
 
     with pytest.raises(Exception) as exception:
@@ -89,9 +85,7 @@ def test_average_wind_speed_with_bad_flow_parameters(amun_session):
         lossesAvailability=0.1,
         lossesWake=0,
         numberOfTurbines=12,
-        roughnessLength=0.02,
-        usePowerCurveSmoothing=False,
-        useReanalysisCorrection=True,
+        usePowerCurveSmoothing=False
     )
 
     # An exception is raised for API v1
@@ -128,9 +122,7 @@ def test_bulk_calculation(amun_session):
         lossesAvailability=0.1,
         lossesWake=0,
         numberOfTurbines=12,
-        roughnessLength=0.02,
-        usePowerCurveSmoothing=False,
-        useReanalysisCorrection=False,
+        usePowerCurveSmoothing=False
     )
 
     results = amun_session.run_load_factors_for_parameters_batch(
@@ -159,9 +151,7 @@ def test_v1_and_v2_calculations(amun_session, snapshot):
         lossesAvailability=0.1,
         lossesWake=0,
         numberOfTurbines=12,
-        roughnessLength=0.02,
-        usePowerCurveSmoothing=False,
-        useReanalysisCorrection=False,
+        usePowerCurveSmoothing=False
     )
 
     # The only thing that changes between usage of v1 and v2 is the version number
@@ -193,9 +183,7 @@ def test_unsupported_version(amun_session):
         lossesAvailability=0.1,
         lossesWake=0,
         numberOfTurbines=12,
-        roughnessLength=0.02,
-        usePowerCurveSmoothing=False,
-        useReanalysisCorrection=False,
+        usePowerCurveSmoothing=False
     )
 
     with pytest.raises(Exception) as exception: