+++ docusaurus-site/docs/changelog.md
@@ -4,8 +4,19 @@ title: Changelog
 description: Whats new in the SDK
 ---
 
+## Version 1.0.5 ![Latest](https://img.shields.io/badge/-Latest-ffcc00?style=for-the-badge)
 
-## Version 1.0.3 ![Latest](https://img.shields.io/badge/-Latest-ffcc00?style=for-the-badge)
+_Date: 11-04-2024_
+
+- Remove unused `roughnessLength` parameter from Load Factor API
+
+## Version 1.0.4
+
+_Date: 2-04-2024_
+
+- `LoadFactorBaseParameters` class no longer provides `useReanalysisCorrection` field which in reality only applies to the `BuiltInWindParameters`
+
+## Version 1.0.3
 
 _Date: 5-02-2024_